feat(lastfm): add first-artist pipeline step for multi-artist tracks
